% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html_tree.R
\name{generate_html_tree}
\alias{generate_html_tree}
\title{Generate HTML Tree for JSON Schema}
\usage{
generate_html_tree(schema_file)
}
\arguments{
\item{schema_file}{Path to the JSON schema file (relative to \code{inst/json_schemas}).}
}
\value{
An HTML representation of the schema.
}
\description{
Creates an HTML tree representation of a JSON schema.
}
