% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_schema_diagram.R
\name{generate_schema_diagram}
\alias{generate_schema_diagram}
\title{Generate Diagram from JSON Schema}
\usage{
generate_schema_diagram(schema_file)
}
\arguments{
\item{schema_file}{Path to the JSON schema file.}
}
\value{
A DiagrammeR graph object.
}
\description{
Generates a graph representation of a JSON schema using the DiagrammeR package.
}
\examples{
schema_file <- system.file("schemas", "person_schema.json", package = "PIFDataModel")
generate_schema_diagram(schema_file)

}
