% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_json_schema.R
\name{load_json_schema}
\alias{load_json_schema}
\title{Load a JSON Schema}
\source{
\url{https://github.com/fsqanalytics/PIFDataModel/blob/main/inst/schemas/user_schema.json}
}
\usage{
load_json_schema(schema_file_name, package = "PIFDataModel")
}
\arguments{
\item{schema_file_name}{A string representing the name of the JSON schema file (e.g., "user_schema.json").}

\item{package}{The name of the package containing the schema. Defaults to \code{"PIFDataModel"}.}
}
\value{
A list representing the schema.
}
\description{
Loads a JSON schema from the package's \verb{inst/schemas/} directory.
}
\examples{
library(PIFDataModel)
my_schema <- load_json_schema(schema_file_name = "user_schema.json")
print(my_schema)
}
