% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_json_schema.R
\name{plot_json_schema}
\alias{plot_json_schema}
\title{Generate HTML Tree for JSON Schema}
\usage{
plot_json_schema(schema_file)
}
\arguments{
\item{schema_file}{A string representing the path to the JSON schema file.}
}
\value{
An HTML widget containing the interactive graph.
}
\description{
This function takes a JSON schema file as input and generates an interactive
HTML tree representation of the schema using a force-directed graph.
}
\examples{
schema_file <- system.file("schemas", "user_schema.json", package = "PIFDataModel")
plot_json_schema(schema_file)

}
