% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_schema_viewer.R
\name{shiny_schema_viewer}
\alias{shiny_schema_viewer}
\title{Run JSON Schema Viewer App with Diagram Feature}
\usage{
shiny_schema_viewer()
}
\description{
Launches a Shiny app to interactively view JSON schemas and generate diagrams.
}
