% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_html_tree.R
\name{generate_html_tree}
\alias{generate_html_tree}
\title{Generate HTML Tree for JSON Schema}
\usage{
generate_html_tree(schema_file)
}
\arguments{
\item{schema_file}{A file path to a JSON schema. If the file is part of a
package, use \code{system.file} to locate it.}
}
\value{
A character string containing the HTML tree.
}
\description{
This function takes a JSON schema file as input and generates an HTML tree
representation of the schema.
}
\examples{
schema_file <- system.file("schemas/user_schema.json", package = "PIFDataModel")
html_representation <- generate_html_tree(schema_file)
cat(html_representation) # Output to console, or write to a file
# write(html_representation, file = "schema_tree.html")

}
