% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_schema_diagram.R
\name{generate_schema_diagram}
\alias{generate_schema_diagram}
\title{Generate Diagram from JSON Schema}
\usage{
generate_schema_diagram(schema_file, debug = FALSE)
}
\arguments{
\item{schema_file}{Path to the JSON schema file.}
}
\value{
A DiagrammeR graph object.
}
\description{
Generates a graph representation of a JSON schema using the DiagrammeR package.
}
\examples{
# Example usage:
# generate_schema_diagram("user_schema.json")
}
