% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_schema.R
\docType{data}
\name{user_schema}
\alias{user_schema}
\title{Example JSON Schema: User Schema}
\format{
A JSON file
}
\source{
: \href{.../schemas/user_schema.json}{View JSON file}
}
\description{
A JSON schema for validating user objects. This schema defines the required fields
and their types for user objects.
}
\examples{
\dontrun{
# Contents of the schema:
cat(readLines(system.file("schemas", "user_schema.json", package = "PIFDataModel")), sep = "\n")
}
}
\keyword{dataset}
