% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_json_schema.R
\name{validate_json_schema}
\alias{validate_json_schema}
\title{Validate JSON Against a Schema}
\usage{
validate_json_schema(schema_file, json_file)
}
\arguments{
\item{schema_file}{Path to the JSON schema file.}

\item{json_file}{Path to the JSON file to validate.}
}
\value{
Logical value indicating if the JSON is valid.
}
\description{
Validates a JSON file against a given JSON schema.
}
\examples{
#schema_file <- system.file("schemas", "bacteria_count_schema.json", package = "PIFDataModel")
#json_file <- system.file("data", "valid_sample_data.json", package = "PIFDataModel")
#validate_json_schema(schema_file, json_file)
#json_file <- system.file("data", "invalid_sample_data.json", package = "PIFDataModel")
#validate_json_schema(schema_file, json_file)
}
